FIX 103
# FIX Ticket
 - #103: As a Backend developer I want to separate document upload from split functionality

**Description**:

- ChatGendoxMessage
  - implement the pass the moderation check message

- CompletionService
  - create new Gendox Exception for not pass the moderation check

- Rename the Sections variable hasContentWarning
  - DocumentInstanceSection
  - V20231107_150400__Add_column_at_document_sections.sql

- Create the DocumentSectionService for sections and secitons metadata service

- ListenerService
  - create the response message if moderation check pass

### Related Ticket: #103
