+++ gendox-core-api/src/main/java/dev/ctrlspace/gendox/gendoxcoreapi/services/EmbeddingService.java
@@ -29,13 +29,13 @@ public class EmbeddingService {
     private EmbeddingRepository embeddingRepository;
     private AuditLogsRepository auditLogsRepository;
     private EmbeddingGroupRepository embeddingGroupRepository;
-    private DocumentService documentService;
     private MessageRepository messageRepository;
     private TypeService typeService;
     private AiModelRepository aiModelRepository;
     private OpenAiEmbeddingConverter openAiEmbeddingConverter;
     private SecurityUtils securityUtils;
     private ProjectAgentService projectAgentService;
+    private DocumentSectionService documentSectionService;
 
 
     @Autowired
@@ -46,24 +46,24 @@ public class EmbeddingService {
                             EmbeddingRepository embeddingRepository,
                             AuditLogsRepository auditLogsRepository,
                             EmbeddingGroupRepository embeddingGroupRepository,
-                            DocumentService documentService,
                             MessageRepository messageRepository,
                             TypeService typeService,
                             OpenAiEmbeddingConverter openAiEmbeddingConverter,
                             AiModelRepository aiModelRepository,
                             SecurityUtils securityUtils,
-                            ProjectAgentService projectAgentService) {
+                            ProjectAgentService projectAgentService,
+                            DocumentSectionService documentSectionService) {
         this.aiModelService = aiModelService;
         this.embeddingRepository = embeddingRepository;
         this.auditLogsRepository = auditLogsRepository;
         this.embeddingGroupRepository = embeddingGroupRepository;
-        this.documentService = documentService;
         this.messageRepository = messageRepository;
         this.typeService = typeService;
         this.aiModelRepository = aiModelRepository;
         this.openAiEmbeddingConverter = openAiEmbeddingConverter;
         this.securityUtils = securityUtils;
         this.projectAgentService = projectAgentService;
+        this.documentSectionService = documentSectionService;
     }
 
     public Embedding createEmbedding(Embedding embedding) throws GendoxException {
@@ -141,7 +141,7 @@ public class EmbeddingService {
         embeddingGroup.setGroupingStrategyType(typeService.getGroupingTypeByName("SIMPLE_SECTION").getId());
         embeddingGroup.setSemanticSearchModelId(aiModelRepository.findByName("Ada2").getId());
 
-        if (message_id!= null){
+        if (message_id != null) {
             embeddingGroup.setMessageId(message_id);
         }
 
@@ -207,7 +207,7 @@ public class EmbeddingService {
 
 
         Set<UUID> nearestEmbeddingsIds = nearestEmbeddings.stream().map(emb -> emb.getId()).collect(Collectors.toSet());
-        List<DocumentInstanceSection> sections = documentService.getSectionsByEmbeddingsIn(projectId, nearestEmbeddingsIds);
+        List<DocumentInstanceSection> sections = documentSectionService.getSectionsByEmbeddingsIn(projectId, nearestEmbeddingsIds);
 
         return sections;
     }
